feat: use default middleware only in tests
